refactor(metrics): extract foundation-pure surface to proximadb-metrics (Slice 1)
First metrics-extraction increment. Creates a new control-layer crate
`proximadb-metrics` and moves the foundation-pure metrics surface (collectors
minus engine/graph, aggregator, exporters, schema, *_metrics types,
UnifiedMetricsCollector) out of the root src/metrics. The storage-coupled
collectors (engine, graph), the persistence layer (store), the updater, the
query_service (uses store), and the consumption/metering/cache/pinning/
tier-migration metrics stay root behind a future MetricsStoragePort.

Path rewrites: crate::metrics:: -> crate:: (self-refs); 2 doc-links to
root modules (crate::query/cluster) reworded. Root src/metrics/mod.rs +
collectors/mod.rs -> re-export shims (glob the clean modules from the crate,
declare the deferred modules, preserve all crate::metrics::* paths). New crate
depends on catalog + foundation (kernel/proto/config/hardware) + chrono.

Two dead-code / false-clean files caught during the move and kept root:
- eventlog_metrics.rs is ORPHAN in root (never declared, never referenced, like
  graph_linking) with a latent .inc_by(u64)-on-f64-counter bug -- moved back to
  root to preserve its status (don't revive dead code in a move slice).
- query_service.rs uses super::store (deferred) -- not actually clean, kept root.

Crate registered in workspace members + workspace.dependencies; root depends on it.
Verification: crate + root clippy --lib --bins -D warnings clean; fmt clean;
layering clean (storage up-edge ratchet unchanged at 197); 75 crate + 156 root
metrics tests pass.
