+++ crates/control/proximadb-metrics/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+name = "proximadb-metrics"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB metrics collection — collectors (access/document/filesystem/query/storage/system), aggregators, exporters, and per-subsystem metric types. Extracted from the root crate (decomposition metrics Slice 1); the storage-coupled collectors stay in the root behind a MetricsStoragePort."
+
+[lib]
+name = "proximadb_metrics"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+dashmap.workspace = true
+# Root-only deps (not in workspace.dependencies) — same specs as the root crate.
+lazy_static = "1.5"
+prometheus = "0.14"
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+chrono.workspace = true
+# Foundation + control deps. Metrics types are keyed by collection (catalog) and
+# read hardware caps; the rest is foundation.
+proximadb-catalog.workspace = true
+proximadb-hardware.workspace = true
+proximadb-kernel.workspace = true
+proximadb-proto.workspace = true
+proximadb-config.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
+
+[lints]
+workspace = true
