+++ crates/control/proximadb-metrics/src/advisor_observations_metrics.rs
