+++ crates/control/proximadb-metrics/src/collectors/access_pattern.rs
