+++ crates/control/proximadb-metrics/src/collectors/document.rs
