+++ crates/control/proximadb-metrics/src/fusion.rs
