+++ crates/control/proximadb-metrics/src/primary_pod_metrics.rs
