+++ crates/control/proximadb-metrics/src/td066_metrics.rs
