+++ crates/control/proximadb-metrics/src/wal_scan_metrics.rs
